rebase on master and packages renaming
