+++ src/commandes/MangaLinkCommand.ts
