+++ src/commandes/TagVoc.ts
@@ -1,23 +1,20 @@
 import { CommandInteraction } from 'discord.js';
 import AbstractCommand, { COMMAND_BEAN_TYPE } from './AbstractCommand.js';
 import TypeHelp from '../entity/typeHelp.js';
-import {
-  AudioPlayerStatus,
-  DiscordGatewayAdapterCreator,
-  joinVoiceChannel,
-} from '@discordjs/voice';
-import EventError from '../errors/EventError.js';
 import CommandInteractionService from '../services/CommandInteractionService.js';
-import ResourcesService from '../services/ResourcesService.js';
 import injector from 'wire-dependency-injection';
 
 export default class TagVocCommand extends AbstractCommand {
-  private commandInteractionService: CommandInteractionService = injector.autoWire(
-    CommandInteractionService,
-    (b) => (this.commandInteractionService = b)
-  );
+  static {
+    injector.instance(this.name, this, {
+      category: COMMAND_BEAN_TYPE,
+      wiring: [CommandInteractionService.name],
+    });
+  }
 
-  public constructor() {
+  public constructor(
+    private commandInteractionService: CommandInteractionService
+  ) {
     super({
       id: 'tagvoc',
       name: ['tagvoc'],
@@ -28,16 +25,14 @@ export default class TagVocCommand extends AbstractCommand {
   }
 
   public async run(commandInteraction: CommandInteraction) {
-    const commandInteractionService = this
-      .commandInteractionService as CommandInteractionService;
-
-    const guildMember = await commandInteractionService.checkDiscordGuildMember(
-      commandInteraction
-    );
+    const guildMember =
+      await this.commandInteractionService.checkDiscordGuildMember(
+        commandInteraction
+      );
 
     const channelId = guildMember.voice.channelId;
     if (!channelId) {
-      commandInteraction.reply({
+      await commandInteraction.reply({
         content: `Vous n'êtes pas dans un channel vocal`,
         ephemeral: true,
       });
@@ -46,7 +41,7 @@ export default class TagVocCommand extends AbstractCommand {
 
     const members = guildMember.voice.channel?.members;
     if (!members) {
-      commandInteraction.reply({
+      await commandInteraction.reply({
         content: `Il n'y a pas d'user dans le channel`,
         ephemeral: true,
       });
@@ -57,16 +52,12 @@ export default class TagVocCommand extends AbstractCommand {
       .map((e) => `<@${e.id}>`)
       .join(' ');
     if (!message) {
-      commandInteraction.reply({
+      await commandInteraction.reply({
         content: `Vous n'avez malheureusement pas d'ami`,
         ephemeral: true,
       });
       return;
     }
-    commandInteraction.reply(message);
+    await commandInteraction.reply(message);
   }
 }
-
-injector.registerBean(TagVocCommand, {
-  type: COMMAND_BEAN_TYPE,
-});
