+++ src/commandes/implementable/ImplementableMangaCommand.ts
